refactor(Types): extract out special ObjectType logic up the stack
Also, we extract out the type_class_mapping and get_type_instance
into a TypeMappings class located in Hub package.
Logic for the special case of ObjectType is extracted up the stack
to the Symbol class.
Overall this is a beautification effort.
