+++ storyscript/compiler/semantics/symbols/Symbols.py
@@ -1,6 +1,10 @@
 # -*- coding: utf-8 -*-
+from storyhub.sdk.service.Output import Output as ServiceOutput
+
 from storyscript.compiler.semantics.types.Indexing import IndexKind
-from storyscript.compiler.semantics.types.Types import BaseType, ObjectType
+from storyscript.compiler.semantics.types.Types import BaseType, MapType, \
+    ObjectType
+from storyscript.hub.Hub import TypeMappings
 
 
 class StorageClass:
@@ -130,10 +134,28 @@ class Symbol:
             assert isinstance(type_, BaseType)
 
         cur_type = symbol.type()
-        if isinstance(cur_type, ObjectType):
-            new_type = cur_type.index(tree, symbol, name, type_, kind)
-        else:
-            new_type = cur_type.index(type_, kind)
+        new_type = cur_type.index(type_, kind)
+        if isinstance(cur_type, ObjectType) and new_type is not None:
+            obj = new_type
+            if isinstance(obj, ServiceOutput):
+                prop = obj.property(name)
+                tree.expect(prop is not None, 'service_output_invalid_prop',
+                            object=symbol.name(), prop=name)
+                new_type = TypeMappings.get_type_instance(var=prop)
+            elif isinstance(obj, dict):
+                # right now we are only storing a dict as the wrapped
+                # value inside a ObjectType instance for the special case
+                # of `app` keyword.
+                assert symbol.name() == 'app'
+                prop = obj.get(name, None)
+                tree.expect(prop is not None, 'object_invalid_prop',
+                            object=symbol.name(), prop=name)
+                # for the app we already store the correct TypeClass in Symbol
+                # type therefore no need to perform mapping.
+                new_type = prop.type()
+            else:
+                assert isinstance(obj, MapType)
+                new_type = obj.index(type_, IndexKind.INDEX)
 
         if kind == IndexKind.DOT:
             tree.expect(new_type is not None,
