+++ tests/e2e/update
@@ -14,7 +14,7 @@ from bom_open import bom_open
 
 from click import unstyle
 
-import storyscript.Hub as StoryHub
+import storyscript.hub.Hub as StoryHub
 from storyscript.Api import Api
 from storyscript.App import _clean_dict
